player: Factor out state into a single State enum
This not only gives a bit more information about what's going on within
a specific player tile, but we now have a single state property which we
can use to boil down all the various states of the subwidgets to the
CasterPlayer class.

Right now we do not yet take into account the QMediaPlayer's state, but
that's something for future improvements.

So the result is now that the user now should no longer be in the dark
if Qt Multimedia fails to load a file.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>