board: Refactor to use a single hash for layout
Having all the keys copy & pasted all over the place is a lot of churn
and doesn't need to be like that.

We're heading on to saving the state of the board layout and duplicating
even more of these lines all over the place is not magically cleaning up
the code.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>