Save and restore layout on exit/startup
We already have a map of all the keys and their corresponding
CasterBoardLayout values, which link to the corresponding widget.

So when closing the window, the layout is saved using QSettings within
the [boards] section in an array mapping from the key id to the file
name, which looks like this:

[boards]
1\49=/path/to/foo.ogg
1\title=aaa
2\69=/path/to/bar.flac
2\title=bbb
size=2

Of course, if we have more complex layouts which doesn't involve a grid,
this will fall short, but we can still implement this once that's the
case.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>
Fixes: #3