CasterBoard: Refactor further + make more readable
The pull request made by @freqlabs is much more readable, because it
also gets rid of the hardcoded grid coordinates.

So this is mainly a rebase of his commit 04f5d1d92c3cc8b2d0312961b4d589
in #9 on top of my changes:

    Use a QMap to simplify CasterBoard

    Instead of using separate member variables for every player, a
    single map can both store the players and provide a convenient
    mechanism for mapping key press events to the appropriate player.

    The map uses a Qt::Key symbol as the key and stores the pointer to
    each widget as the value.

    To populate the map, we iterate over each character in a string,
    where the characters correspond to the labels of the hotkey widgets.
    The corresponding Qt::Key values are stored in another array that is
    accessed in parallel. A counter keeps track of the offset in the
    keys array and is also used to calculate the grid position of each
    widget.

Instead of a map I'm still using a QHash, because it doesn't allow for
duplicates and is also faster in lookup.

I've also noticed that he's been using C++11 constructs, so I'm also
using these for saveLayout() and restoreLayout() as well.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>
Cc: @freqlabs