Add Issue Template and Guidelines
Adds a template for the issues tracker to keep things more organized and adds a Contributing Guidelines for Reporting Bug Reports or Requesting Features. We will probably want to expand the Contributor's Guidelines to incude code style and info for beginners. Also probably add a Contributors Code of Conduct in the Future.